Enable contacts, receive updated preferences from household, neighbours and parents
Enable setting up contacts in the setup, and daily receive updated preferences from your own household, some neighbours and some other parents.

Adds two sliders, one with the average amount of daily-neighbour-contacts and one with the average of daily-parent-contacts.

TODO: Update contacts when household move out the neighbourhood. Currently they are removed from the model (die) but not from the lists of contacts.