OVERHAUL: Initialize cars and subscriptions based on preferences
Overhaul the setup-cars function (renamed setup-cars-subscriptions) to setup cars and add subscriptions based on the preferences of residents compared to the other residents.

The problem encountered was that we both wanted to initialize a fixed number of cars in the neighbourhood, because we know that number approximately and want to experiment with it, but also wanted to give the residents with the highest car preferences cars, and not residents with low residents.

The solution was using the NetLogo Stats Extension (https://github.com/cstaelin/Stats-Extension) to calculate the quantiles of the current preferences for each modality. If a resident has a higher preference than the nth quantile (which can be varied with a slider), he gets a car or subscription for that modality.

This makes it very robust to changing other parameters and also works well with new households moving in.

Adds sliders for initial-shared-car-subscriptions (5%), initial-public-transport-subscription (20%), initial-car-chance-parent (65%) and initial-car-chance-child (20%).